feat(ui): add copy path, toolbar toggle, and breadcrumb view controls
- Right-click on sidebar items now shows Copy Path and Copy Relative Path
- Added hide/show titlebar toggle button in breadcrumb bar
- Moved reading mode and theme buttons from titlebar to breadcrumb
- Extracted ThemeButton as a standalone reusable component
- Added visual separator (border-left) between view controls and file actions
- Fixed MSVC build environment via .cargo/config.toml
