+++ src/components/chrome/breadcrumb.tsx
@@ -1,6 +1,20 @@
-import { Check, ChevronRight, Copy, FileDown, FilePlus2, FileText, FolderOpen, PanelLeftClose, PanelLeftOpen } from "lucide-react";
+import {
+  BookOpen,
+  Check,
+  ChevronRight,
+  Copy,
+  FileDown,
+  FilePlus2,
+  FileText,
+  FolderOpen,
+  PanelLeftClose,
+  PanelLeftOpen,
+  PanelTopClose,
+  PanelTopOpen,
+} from "lucide-react";
 import { Button, Icon } from "@/components/primitives";
 import { shortcutLabel, startWindowDrag, useI18n, type Translate } from "@/lib";
+import { ThemeButton } from "./theme-button";
 import exciteUrl from "@/assets/mascot/excite.png";
 
 export type SaveStatus = "idle" | "dirty" | "saving" | "saved";
@@ -17,6 +31,12 @@ type BreadcrumbProps = {
   onCopyMarkdown?: () => void;
   onExportPdf?: () => void;
   copyPulse?: boolean;
+  titlebarVisible: boolean;
+  onToggleTitlebar: () => void;
+  readingMode: boolean;
+  onToggleReading: () => void;
+  vimOn?: boolean;
+  onToggleVim?: () => void;
 };
 
 const MAX_SEGMENTS = 4;
@@ -52,6 +72,12 @@ export function Breadcrumb({
   onCopyMarkdown,
   onExportPdf,
   copyPulse = false,
+  titlebarVisible,
+  onToggleTitlebar,
+  readingMode,
+  onToggleReading,
+  vimOn,
+  onToggleVim,
 }: BreadcrumbProps) {
   const { t } = useI18n();
   const path = activePath ?? rootPath;
@@ -80,16 +106,9 @@ export function Breadcrumb({
           segments.map((seg, i) => (
             <span key={`${seg}-${i}`} className="mdv-breadcrumb__seg-row">
               {i > 0 ? (
-                <Icon
-                  icon={ChevronRight}
-                  size={11}
-                  strokeWidth={1.5}
-                  title="separator"
-                />
+                <Icon icon={ChevronRight} size={11} strokeWidth={1.5} title="separator" />
               ) : null}
-              <span
-                className={`mdv-breadcrumb__seg${i === segments.length - 1 ? " is-leaf" : ""}`}
-              >
+              <span className={`mdv-breadcrumb__seg${i === segments.length - 1 ? " is-leaf" : ""}`}>
                 {seg}
               </span>
             </span>
@@ -119,46 +138,72 @@ export function Breadcrumb({
       </div>
 
       <div className="mdv-breadcrumb__actions" data-tauri-drag-region>
-        {onCopyMarkdown ? (
-          <button
-            type="button"
-            className={`mdv-copybtn${copyPulse ? " is-copied" : ""}`}
-            data-tooltip={copyPulse ? t("app.copied") : shortcutLabel(t("app.copyMarkdownShortcut"))}
-            aria-label={copyPulse ? t("app.copied") : t("app.copyMarkdown")}
-            onClick={onCopyMarkdown}
-          >
-            <span className="mdv-copybtn__icon mdv-copybtn__icon--copy" aria-hidden>
-              <Icon icon={Copy} size={12} strokeWidth={1.5} />
-            </span>
-            <span className="mdv-copybtn__icon mdv-copybtn__icon--check" aria-hidden>
-              <Icon icon={Check} size={13} strokeWidth={2} />
-            </span>
-          </button>
-        ) : null}
+        {/* view controls: hide titlebar, reading mode, theme */}
         <Button
-          data-tooltip={shortcutLabel(t("app.exportPdfShortcut"))}
-          aria-label={t("app.exportPdf")}
-          onClick={onExportPdf}
-          icon={<Icon icon={FileDown} size={13} strokeWidth={1.5} />}
+          data-tooltip={titlebarVisible ? t("title.hideBreadcrumb") : t("title.showBreadcrumb")}
+          aria-label={titlebarVisible ? t("title.hideBreadcrumb") : t("title.showBreadcrumb")}
+          aria-pressed={!titlebarVisible}
+          onClick={onToggleTitlebar}
+          icon={
+            <Icon
+              icon={titlebarVisible ? PanelTopClose : PanelTopOpen}
+              size={14}
+              strokeWidth={1.5}
+            />
+          }
         />
         <Button
-          data-tooltip={shortcutLabel(t("app.newFileShortcut"))}
-          aria-label={t("app.newFile")}
-          onClick={onNewFile}
-          icon={<Icon icon={FilePlus2} size={13} strokeWidth={1.5} />}
+          data-tooltip={readingMode ? t("title.exitReadingTooltip") : shortcutLabel(t("title.readingModeShortcut"))}
+          aria-label={readingMode ? t("title.exitReading") : t("title.readingMode")}
+          aria-pressed={readingMode}
+          onClick={onToggleReading}
+          icon={<Icon icon={readingMode ? BookOpen : BookOpen} size={14} strokeWidth={1.5} />}
         />
-        <Button
-          data-tooltip={shortcutLabel(t("app.openFileShortcut"))}
-          aria-label={t("app.openFile")}
-          onClick={onOpenFile}
-          icon={<Icon icon={FileText} size={13} strokeWidth={1.5} />}
-        />
-        <Button
-          data-tooltip={shortcutLabel(t("app.openFolderShortcut"))}
-          aria-label={t("app.openFolder")}
-          onClick={onOpenFolder}
-          icon={<Icon icon={FolderOpen} size={13} strokeWidth={1.5} />}
-        />
+        <ThemeButton vimOn={vimOn} onToggleVim={onToggleVim} />
+
+        {/* file actions — border-left matches the status→actions separator */}
+        <div className="mdv-breadcrumb__file-actions">
+          {onCopyMarkdown ? (
+            <button
+              type="button"
+              className={`mdv-copybtn${copyPulse ? " is-copied" : ""}`}
+              data-tooltip={copyPulse ? t("app.copied") : shortcutLabel(t("app.copyMarkdownShortcut"))}
+              aria-label={copyPulse ? t("app.copied") : t("app.copyMarkdown")}
+              onClick={onCopyMarkdown}
+            >
+              <span className="mdv-copybtn__icon mdv-copybtn__icon--copy" aria-hidden>
+                <Icon icon={Copy} size={12} strokeWidth={1.5} />
+              </span>
+              <span className="mdv-copybtn__icon mdv-copybtn__icon--check" aria-hidden>
+                <Icon icon={Check} size={13} strokeWidth={2} />
+              </span>
+            </button>
+          ) : null}
+          <Button
+            data-tooltip={shortcutLabel(t("app.exportPdfShortcut"))}
+            aria-label={t("app.exportPdf")}
+            onClick={onExportPdf}
+            icon={<Icon icon={FileDown} size={13} strokeWidth={1.5} />}
+          />
+          <Button
+            data-tooltip={shortcutLabel(t("app.newFileShortcut"))}
+            aria-label={t("app.newFile")}
+            onClick={onNewFile}
+            icon={<Icon icon={FilePlus2} size={13} strokeWidth={1.5} />}
+          />
+          <Button
+            data-tooltip={shortcutLabel(t("app.openFileShortcut"))}
+            aria-label={t("app.openFile")}
+            onClick={onOpenFile}
+            icon={<Icon icon={FileText} size={13} strokeWidth={1.5} />}
+          />
+          <Button
+            data-tooltip={shortcutLabel(t("app.openFolderShortcut"))}
+            aria-label={t("app.openFolder")}
+            onClick={onOpenFolder}
+            icon={<Icon icon={FolderOpen} size={13} strokeWidth={1.5} />}
+          />
+        </div>
       </div>
     </div>
   );
