+++ src/components/chrome/index.ts
@@ -2,3 +2,4 @@ export { Logo } from "./logo";
 export { TitleBar } from "./title-bar";
 export { Breadcrumb, type SaveStatus } from "./breadcrumb";
 export { StatusBar, type VimMode } from "./status-bar";
+export { ThemeButton } from "./theme-button";
