+++ src/locales/en.json
@@ -49,6 +49,8 @@
   "sidebar.clearContext": "clear context",
   "sidebar.resize": "resize sidebar",
   "sidebar.tokens": "{tokens} tok",
+  "title.hideBreadcrumb": "hide toolbar",
+  "title.showBreadcrumb": "show toolbar",
   "title.unsavedChanges": "unsaved changes",
   "title.exitReading": "exit reading mode",
   "title.exitReadingTooltip": "exit reading (esc)",
@@ -66,6 +68,9 @@
   "theme.group.catppuccin": "catppuccin",
   "theme.group.ai": "ai",
   "theme.group.crafted": "crafted",
+  "menu.copyPath": "copy path",
+  "menu.copyRelativePath": "copy relative path",
+  "menu.pathCopied": "path copied",
   "menu.rename": "rename",
   "menu.newFile": "new file",
   "menu.newFolder": "new folder",
