+++ src/locales/ko.json
@@ -49,6 +49,8 @@
   "sidebar.clearContext": "컨텍스트 지우기",
   "sidebar.resize": "사이드바 크기 조절",
   "sidebar.tokens": "{tokens} tok",
+  "title.hideBreadcrumb": "툴바 숨기기",
+  "title.showBreadcrumb": "툴바 보이기",
   "title.unsavedChanges": "저장하지 않은 변경",
   "title.exitReading": "읽기 모드 나가기",
   "title.exitReadingTooltip": "읽기 나가기 (esc)",
@@ -66,6 +68,9 @@
   "theme.group.catppuccin": "catppuccin",
   "theme.group.ai": "AI",
   "theme.group.crafted": "crafted",
+  "menu.copyPath": "경로 복사",
+  "menu.copyRelativePath": "상대 경로 복사",
+  "menu.pathCopied": "경로가 복사됨",
   "menu.rename": "이름 바꾸기",
   "menu.newFile": "새 파일",
   "menu.newFolder": "새 폴더",
