+++ src/styles/chrome/breadcrumb.css
@@ -112,6 +112,20 @@
   padding: 0 5px;
 }
 
+.mdv-breadcrumb__file-actions {
+  display: inline-flex;
+  align-items: center;
+  gap: 1px;
+  padding-left: 8px;
+  margin-left: 6px;
+  border-left: 1px solid var(--border);
+}
+
+.mdv-breadcrumb__file-actions .mdv-btn {
+  height: 22px;
+  padding: 0 5px;
+}
+
 /* copy-markdown button — crossfade between Copy and Check icons on success */
 .mdv-copybtn {
   position: relative;
